Pre-training_on_the_SUN_data_set/requirements.txt
datasets~=2.8.0
pytorch-lightning
scikit-learn~=1.2.0
transformers~=4.25.1
hugsvision~=0.75.5
piexif~=1.1.3

#warnings
pandas~=1.5.3
numpy~=1.24.1
opencv-python~=4.7.0.68
torch~=1.13.1
torchvision~=0.14.1
seaborn~=0.12.2
matplotlib~=3.6.3
#time
PILLOW~=9.4.0
sklearn~=0.0.post1

torchmetrics~=0.11.0
tensorboard